Minor cleanup of libdl JS functions. NFC.
Split out from a larger refactor I'm working on.

By delaying the creation of the `__self__` handle we can set the
module to be the actual wasm module (Module['asm']) so which avoids
the need for special handling of the main module in `dlsym`.

Remove from `else after return` usage, and avoid looking up propoerties
twice.

Remove `cleanups` property since it was not used anywhere.